third_party/WebKit/LayoutTests/fast/dom/Element/classlist-empty-string.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<script>
debug("Check the classlist length attribute for empty strings");
var div = document.createElement('div');
div.className = '   ';
shouldBe("div.className.length", "3");

div.className = '     ';
shouldBe("div.className.length", "5");
</script>
</html>
